+++ near/cogs/crypto.py
@@ -1,9 +1,9 @@
 import aiohttp
-from bs4 import BeautifulSoup
 
 import discord
 from discord import app_commands
 from discord.ext import commands
+from bs4 import BeautifulSoup
 
 from near.utils import embeds