+++ near/cogs/tools.py
@@ -1,10 +1,12 @@
 import os
 import secrets
 import string
+import shutil
+import glob
 import textwrap
 import urllib
-
 import aiohttp
+
 import discord
 import instaloader
 import privatebinapi
@@ -12,8 +14,9 @@ from discord import app_commands
 from discord.ext import commands
 from zxcvbn import zxcvbn
 
-from near.database import get_embeds
-from near.utils import input_sanitization, errors
+from near.utils import embeds
+from near.utils import errors
+from near.utils import input_sanitization
 
 
 class Tools(commands.Cog):
@@ -23,129 +26,77 @@ class Tools(commands.Cog):
         with open(os.path.join('near', 'assets', 'tenmilpwds.txt'), 'r') as _file:
             self.lines = _file.readlines()
 
-    @app_commands.command(name="lyrics", description="Search the Lyrics of any Song")
-    @app_commands.describe(query="Name of the Song")
-    async def lyrics(self, intercation: discord.Interaction, query: str = None):
-        # Another Option: https://github.com/elmoiv/azapi
-        try:
-            if input_sanitization.check_input(query):
-                if not query:
-                    embed = discord.Embed(title="No search argument!", description="You havent entered anything, so i couldnt find lyrics!", color=get_embeds.Common.COLOR)
-                    embed.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                    embed.set_footer(text=f"Requested by {intercation.user.name}")
-                    return await intercation.response.send_message(embed=embed)
-
-                song = urllib.parse.quote(query)
-
-                async with aiohttp.ClientSession() as lyricsSession:
-                    async with lyricsSession.get(f'https://some-random-api.ml/lyrics?title={song}') as jsondata:
-                        if not 300 > jsondata.status >= 200:
-                            return await intercation.response.send_message(f'Recieved poor status code of {jsondata.status}')
-                        lyricsData = await jsondata.json()
-
-                error = lyricsData.get('error')
-                if error:
-                    return await intercation.response.send_message(f'Recieved unexpected error: {error}')
-
-                songLyrics = lyricsData['lyrics']
-                songArtist = lyricsData['author']
-                songTitle = lyricsData['title']
-                songThumbnail = lyricsData['thumbnail']['genius']
-
-                for chunk in textwrap.wrap(songLyrics, 4096, replace_whitespace=False):
-                    embed = discord.Embed(
-                        title=f'{songTitle} - {songArtist}',
-                        description=chunk,
-                        color=get_embeds.Common.COLOR
-                    )
-                    embed.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                    embed.set_thumbnail(url=songThumbnail)
-                    embed.set_footer(text=f"Requested by {intercation.user.name}")
-                    await intercation.response.send_message(embed=embed)
-            else:
-                raise errors.IllegalInput
-
-        except Exception as e:
-            embed3 = discord.Embed(title=get_embeds.ErrorEmbeds.TITLE, description=get_embeds.ErrorEmbeds.DESCRIPTION, color=get_embeds.ErrorEmbeds.COLOR)
-            embed3.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-            embed3.set_thumbnail(url=get_embeds.ErrorEmbeds.THUMBNAIL)
-            embed3.add_field(name="Error:", value=f"{e}", inline=False)
-            embed3.set_footer(text=f"Requested by {intercation.user.name}")
-            await intercation.response.send_message(embed=embed3)
-
     @app_commands.command(name='passwordgen', description="Generate a very secure and unique password")
     @app_commands.describe(length="Length of the Password to generate.")
     async def passwordgen(self, interaction: discord.Interaction, length: int):
-        # input sanitization not needed here
         try:
 
             if int(length) < 101:
                 c = ''.join((secrets.choice(string.ascii_letters + string.digits + string.punctuation) for i in range(int(length))))
 
-                embed = discord.Embed(title="Password Generator", color=get_embeds.Common.COLOR)
-                embed.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                embed.set_thumbnail(url="https://cdn.discordapp.com/attachments/877796755234783273/880031728369016832/704187.png")
+                embed = embeds.Common(
+                    client=self.client,
+                    interaction=interaction,
+                    title="Password Generator",
+                    thumbnail="https://cdn.discordapp.com/attachments/877796755234783273/880031728369016832/704187.png"
+                )
                 embed.add_field(name="Password Length", value=f"{length}", inline=False)
                 embed.add_field(name="Password", value=f"```{c}```", inline=False)
-                embed.set_footer(text=f"Requested by {interaction.user.name}")
                 await interaction.response.send_message(embed=embed)
 
             else:
-                embed = discord.Embed(title="Password Generator", description="An Error has occured!", color=get_embeds.Common.COLOR)
-                embed.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                embed.set_thumbnail(url="https://cdn.discordapp.com/attachments/877796755234783273/880031728369016832/704187.png")
-                embed.add_field(name="Error", value="The value of the number is high", inline=False)
-                embed.add_field(name="Possible Fix", value="Enter a value below 40", inline=False)
-                embed.set_footer(text=f"Requested by {interaction.user.name}")
-                await interaction.response.send_message(embed=embed)
+                await interaction.response.send_message(embed=embeds.Error(interaction=interaction, client=self.client, error_message=f"The value of the number is high.\nPlease Enter a value below 40,"), ephemeral=False)
 
         except Exception as e:
-            embed3 = discord.Embed(title=get_embeds.ErrorEmbeds.TITLE, description=get_embeds.ErrorEmbeds.DESCRIPTION, color=get_embeds.ErrorEmbeds.COLOR)
-            embed3.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-            embed3.set_thumbnail(url=get_embeds.ErrorEmbeds.THUMBNAIL)
-            embed3.add_field(name=get_embeds.ErrorEmbeds.FIELD_NAME, value=f"{e}", inline=False)
-            embed3.set_footer(text=f"Requested by {interaction.user.name}")
-            await interaction.response.send_message(embed=embed3)
+            await interaction.response.send_message(embed=embeds.Error(interaction=interaction, client=self.client, error_message=f"{e}"), ephemeral=False)
 
     @app_commands.command(name="insta", description="Grab the Instagram Profile Picture of a Profile")
     @app_commands.describe(username="Instagram Profile's Username")
     async def insta(self, interaction: discord.Interaction, username: str):
         try:
-            if input_sanitization.is_instagram_username(username):
-                igpfp = instaloader.Instaloader()
-                igpfp.download_profile(username, profile_pic_only=True)
-                os.chdir(f'{username}')
-                try:
-                    os.system("mv *.jpg ..")
-                except:
-                    os.system("move *.jpg ..")
-                os.chdir("..")
-                try:
-                    os.system("mv *.jpg igtemp.jpg")
-                except:
-                    os.system("ren *.jpg igtemp.jpg")
-                try:
-                    os.system(f'rm -r {username}')
-                except:
-                    os.system(f'DEL {username} /F/Q/S')
-
-                file = discord.File(f'igtemp.jpg', filename="image.jpg")
-                embed = discord.Embed(title="Instagram Profile Picture", description=f"of {username}", color=get_embeds.Common.COLOR)
-                embed.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                embed.add_field(name="Link", value=f"https://instagram.com/{username}", inline=False)
-                embed.set_image(url="attachment://image.jpg")
-                embed.set_footer(text=f"Requested by {interaction.user.name}")
-                await interaction.response.send_message(file=file, embed=embed)
-
-                try:
-                    os.system(f"rm igtemp.jpg")
-                except:
-                    os.remove(f'{username}')
-            else:
-                raise errors.IllegalInput
+            igpfp = instaloader.Instaloader()
+            igpfp.download_profile(username, profile_pic_only=True)
+                
+            os.chdir(f'{username}')
+            jpg_files = glob.glob("*.jpg")
+            for jpg_file in jpg_files:
+                shutil.move(jpg_file, os.path.join("..", jpg_file))
+            os.chdir("..")
+            jpg_files = glob.glob("*.jpg")
+            for jpg_file in jpg_files:
+                shutil.move(jpg_file, "igtemp.jpg")
+                break
+                
+            file = discord.File(f'igtemp.jpg', filename="image.jpg")
+            
+            # Use this if you want to send the image as an embed
+            # ---
+            
+            embed = embeds.Common(
+                client=self.client,
+                interaction=interaction,
+                title="Instagram Profile Picture",
+                description=f"of {username}",
+                thumbnail="https://cdn.discordapp.com/attachments/877796755234783273/880031728369016832/704187.png"
+            )
+            embed.add_field(name="Link", value=f"https://instagram.com/{username}", inline=False)
+            embed.set_image(url="attachment://image.jpg")
+            await interaction.response.send_message(file=file, embed=embed)
+            
+            # Use this if you only want to send the file without an embed
+            # ---
+            # await interaction.response.send_message(file=file)
 
         except Exception as e:
-            await interaction.response.send_message(f"Error: {e}")
+            await interaction.response.send_message(embed=embeds.Error(interaction=interaction, client=self.client, error_message=f"{e}"), ephemeral=False)
+            
+        finally:
+            try:
+                shutil.rmtree(username)
+                os.remove(f'igtemp.jpg')
+            except Exception as e:
+                print(f"Error removing directory {username}: {e}")
+            
 
     @app_commands.command(name="bin", description="Create a PrivateBin from a Text")
     @app_commands.describe(text="Text to be included in the PrivateBin")
@@ -153,26 +104,23 @@ class Tools(commands.Cog):
         try:
             if input_sanitization.check_input(text):
                 privbin = privatebinapi.send("https://bin.teamsds.net/", text=text)
-
-                embed = discord.Embed(title="TeamSDS's PrivateBin", color=get_embeds.Common.COLOR)
-                embed.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                embed.set_thumbnail(url="https://cdn.discordapp.com/attachments/877796755234783273/879586340520480768/large.png")
+                
+                embed = embeds.Common(
+                    client=self.client,
+                    interaction=interaction,
+                    title="TeamSDS's PrivateBin",
+                    thumbnail="https://cdn.discordapp.com/attachments/877796755234783273/879586340520480768/large.png"
+                )
                 embed.add_field(name="ID", value=f"{privbin['id']}", inline=False)
                 embed.add_field(name="URL", value=f"{privbin['full_url']}", inline=False)
                 embed.add_field(name="Passcode", value=f"{privbin['passcode']}", inline=False)
                 embed.add_field(name=f"Text by {interaction.user.name}", value=f"{text}", inline=False)
-                embed.set_footer(text=f"Requested by {interaction.user.name}")
                 await interaction.response.send_message(embed=embed)
             else:
                 raise errors.IllegalInput
 
         except Exception as e:
-            embed3 = discord.Embed(title=get_embeds.ErrorEmbeds.TITLE, description=get_embeds.ErrorEmbeds.DESCRIPTION, color=get_embeds.ErrorEmbeds.COLOR)
-            embed3.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-            embed3.set_thumbnail(url=get_embeds.ErrorEmbeds.THUMBNAIL)
-            embed3.add_field(name=get_embeds.ErrorEmbeds.FIELD_NAME, value=f"{e}", inline=False)
-            embed3.set_footer(text=f"Requested by {interaction.user.name}")
-            await interaction.response.send_message(embed=embed3, ephemeral=True)
+            await interaction.response.send_message(embed=embeds.Error(interaction=interaction, client=self.client, error_message=f"{e}"), ephemeral=False)
 
     @app_commands.command(name="passwordcheck", description="Password Strength Check and Profiler")
     @app_commands.describe(password="Password to analyze")
@@ -180,77 +128,88 @@ class Tools(commands.Cog):
         try:
             if input_sanitization.password_check(password):
                 results = zxcvbn(f"{password}")
-                embed3 = discord.Embed(title="Password Check", description="using Low Budget Password Strength Estimation", color=get_embeds.Common.COLOR)
-                embed3.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-                embed3.set_thumbnail(url="https://iconape.com/wp-content/png_logo_vector/password.png")
+                
+                embed = embeds.Common(
+                    client=self.client,
+                    interaction=interaction,
+                    title="Password Check",
+                    description="using Low Budget Password Strength Estimation (zxcvbn)",
+                    thumbnail="https://iconape.com/wp-content/png_logo_vector/password.png"
+                )
 
-                embed3.add_field(name="Password", value=f"{results['password']}", inline=False)
+                embed.add_field(name="Password", value=f"{results.get('password', '')}", inline=False)
+                embed.add_field(name="Score", value=f"{results['score']}", inline=False)
+                embed.add_field(name="Guesses", value=f"Decimal - {results['guesses']}\nLog10 - {results['guesses_log10']}", inline=False)
+                
+                sequence_info = "\n".join(f"{k} - {v}" for dic in results["sequence"] for k, v in dic.items())
+                embed.add_field(name="Sequence Info", value=sequence_info, inline=False)
 
-                embed3.add_field(name="Score", value=f"{results['score']}", inline=False)
+                embed.add_field(name="Calculation Time", value=f"{results.get('calc_time', '')}", inline=False)
+                crack_times = results.get('crack_times_display', {})
+                embed.add_field(name="Crack Time", value=(
+                    f"Online throttling 100 per hour - {crack_times.get('online_throttling_100_per_hour', '')}\n"
+                    f"Online throttling 10 per second - {crack_times.get('online_no_throttling_10_per_second', '')}\n"
+                    f"Offline slow hashing 1e4 per second - {crack_times.get('offline_slow_hashing_1e4_per_second', '')}\n"
+                    f"Offline fast hashing 1e10 per second - {crack_times.get('offline_fast_hashing_1e10_per_second', '')}"
+                ), inline=False)
 
-                embed3.add_field(name="Guesses", value=f"Decimal - {results['guesses']}\nLog10 - {results['guesses_log10']}", inline=False)
-                try:
-                    embed3.add_field(name="Password", value=f"{results['password']}", inline=False)
-                except:
-                    pass
-
-                sequence_info = ""
-                for dic in results["sequence"]:
-                    for k, v in dic.items():
-                        sequence_info += f"{k} - {v}\n"
-                    sequence_info += "\n"
-
-                embed3.add_field(name="Sequence Info", value=f"{sequence_info}", inline=False)
-
-                try:
-                    embed3.add_field(name="Calculation Time", value=f"{results['calc_time']}")
-                except:
-                    pass
-
-                try:
-                    embed3.add_field(name="Crack Time", value=f"Online throttling 100 per hour - {results['crack_times_display']['online_throttling_100_per_hour']}\nOnline throttling 10 per second - {results['crack_times_display']['online_no_throttling_10_per_second']}\nOffline slow hasing 1e4 per second - {results['crack_times_display']['offline_slow_hashing_1e4_per_second']}\nOffline fast hasing 1e10 per second - {results['crack_times_display']['offline_fast_hashing_1e10_per_second']}", inline=False)
-                except:
-                    pass
-
-                try:
-                    if results['feedback']['warning'] == '':
-                        pass
-                    else:
-                        embed3.add_field(name="Warning", value=f"{results['feedback']['warning']}", inline=False)
-                except:
-                    pass
-
-                try:
-                    if len(results["feedback"]["suggestions"]) == 0:
-                        pass
-                    else:
-                        suggestions_info = ""
-                        for item in results["feedback"]["suggestions"]:
-                            suggestions_info += f"{item}\n"
-                        try:
-                            embed3.add_field(name="Suggestions", value=f"{suggestions_info}", inline=False)
-                        except Exception as e:
-                            print(e)
-                except:
-                    pass
-
+                feedback = results.get('feedback', {})
+                warning = feedback.get('warning', '')
+                if warning:
+                    embed.add_field(name="Warning", value=warning, inline=False)
+                
+                suggestions = "\n".join(feedback.get('suggestions', []))
+                if suggestions:
+                    embed.add_field(name="Suggestions", value=suggestions, inline=False)
+                
                 if f'{password}\n' in self.lines:
-                    embed3.add_field(name="Common Password", value="This password is among the worlds top 1 million most common passwords", inline=False)
+                    embed.add_field(name="Common Password", value="This password is among the worlds top 1 million most common passwords", inline=False)
 
-                embed3.set_footer(text=f"Requested by {interaction.user.name}")
-
-                await interaction.response.send_message(embed=embed3)
+                await interaction.response.send_message(embed=embed)
             else:
                 raise errors.IllegalInput
 
         except Exception as e:
-            embed3 = discord.Embed(title=get_embeds.ErrorEmbeds.TITLE, description=get_embeds.ErrorEmbeds.DESCRIPTION, color=get_embeds.ErrorEmbeds.COLOR)
-            embed3.set_author(name=f"{self.client.user.name}", icon_url=f"{self.client.user.avatar.url}")
-            embed3.set_thumbnail(url=get_embeds.ErrorEmbeds.THUMBNAIL)
-            embed3.add_field(name="Error:", value=f"{e}", inline=False)
-            embed3.set_footer(text=f"Requested by {interaction.user.name}")
-            await interaction.response.send_message(embed=embed3)
+            await interaction.response.send_message(embed=embeds.Error(interaction=interaction, client=self.client, error_message=f"{e}"), ephemeral=False)
 
+    @app_commands.command(name="lyrics", description="Search the Lyrics of any Song")
+    @app_commands.describe(query="Name of the Song")
+    async def lyrics(self, interaction: discord.Interaction, query: str):
+        try:
+            if input_sanitization.check_input(query):
+
+                song = urllib.parse.quote(query)
+
+                async with aiohttp.ClientSession() as lyricsSession:
+                    async with lyricsSession.get(f'https://some-random-api.com/others/lyrics?title={song}') as jsondata:
+                        if not 300 > jsondata.status >= 200:
+                            return await interaction.response.send_message(f'Recieved poor status code of {jsondata.status}')
+                        lyricsData = await jsondata.json()
+
+                error = lyricsData.get('error')
+                if error:
+                    return await interaction.response.send_message(f'Recieved unexpected error: {error}')
+
+                songLyrics = lyricsData['lyrics']
+                songArtist = lyricsData['author']
+                songTitle = lyricsData['title']
+                songThumbnail = lyricsData['thumbnail']['genius']
+
+                for chunk in textwrap.wrap(songLyrics, 4096, replace_whitespace=False):
+                    embed = embeds.Common(
+                        client=self.client,
+                        interaction=interaction,
+                        title=f'{songTitle} - {songArtist}',
+                        description=chunk,
+                        thumbnail="https://iconape.com/wp-content/png_logo_vector/password.png"
+                    )
+                    embed.set_thumbnail(url=songThumbnail)
+                    await interaction.response.send_message(embed=embed)
+            else:
+                raise errors.IllegalInput
+
+        except Exception as e:
+            await interaction.response.send_message(embed=embeds.Error(interaction=interaction, client=self.client, error_message=f"{e}"), ephemeral=False)
 
 async def setup(client: commands.Bot):
     await client.add_cog(Tools(client))